Kodi's Code
Kodi's Code:

was impossible to find the coral this is a major improvement and it was failing to take the prop off your back. Added coral props and drew an outline on them when within distance. Put distance check loop in a circle zone where the blip spawns so the loop doesnt fire until you are within that zone.

place on ground and freeze entity

add Entity ids to table