Fix sell reward, rework modifier
As noted in issue #89 selling a large quantity of coral resulted in a $1 reward.

This commit fixes the cash payout, and reworks the old modifier into a bonus system.
Prior to these changes, the more coral you sold, the less money would be rewarded.
This changes it to give a bonus percentage that gets higher when selling more coral, giving reward for risking having more illegal items in your pockets.